chore: combine Utils files
